+++ Toolkit/RSPiX/GREEN/Task/task.h
@@ -38,7 +38,7 @@ class RTask
 
 		// Typedef to override this RTask's timer.  If overridden, this RTask
 		// will use this callback to get the time.
-		typedef int32_t (*TimeFunc)(	// Returns time as a long in milliseconds.
+      typedef int32_t (*TimeFunc)(	// Returns time as a int32_t in milliseconds.
 			int32_t lTimeUser);			// User defined m_lTimeUser.
 
 	/////////////////////// Con/Destruction ////////////////////////////////////
@@ -57,7 +57,7 @@ class RTask
 		int16_t IsActive(void) { return m_sActive; }
 
 		// Returns the current time based on either the user base or Blue.
-		long GetTime(void)
+      int32_t GetTime(void)
 			{ return (m_fnTime == nullptr ? rspGetMilliseconds() : (*m_fnTime)(m_lTimeUser)); }
 		
 	////////////////////////// Methods ////////////////////////////////////////
@@ -98,8 +98,8 @@ class RTask
 		TaskFunc		m_fnTask;		// User specified task function.
 		uint32_t			m_ulUser;		// User specified parm to task function.
 
-		long			m_lInterval;	// User specified timer interval.
-		long			m_lNextExpiration;	// Next time to call task.
+      int32_t			m_lInterval;	// User specified timer interval.
+      int32_t			m_lNextExpiration;	// Next time to call task.
 
 	protected:
 		// The following member variables are NOT safe to tamper with from 
@@ -107,7 +107,7 @@ class RTask
 
 		int16_t				m_sActive;	// TRUE if active (in list), FALSE otherwise.
 		TimeFunc			m_fnTime;	// Custom time function.
-		long				m_lTimeUser;// Custom time function user value.
+      int32_t				m_lTimeUser;// Custom time function user value.
 
 
 		/////////////////////// Static members /////////////////////////////////
