+++ Toolkit/RSPiX/ORANGE/Channel/channel.cpp
@@ -40,7 +40,7 @@
 // Usefull test code to make sure it will really compile.  If templated stuff
 // isn't actually instantiated, the compiler will miss many (or all) errors.
 RChannel<int16_t> test;
-RChannel<long> test2(RChannel_ArrayOfPtrs);
+RChannel<int32_t> test2(RChannel_ArrayOfPtrs);
 
 
 ////////////////////////////////////////////////////////////////////////////////
